Added docs and refactored
